+++ internal/metrics/cpu/parser.go
@@ -3,8 +3,6 @@ package cpu
 import (
 	"context"
 	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/sitnikovik/sysmon/internal/metrics/utils"
 	"github.com/sitnikovik/sysmon/internal/metrics/utils/cmd"
@@ -68,31 +66,6 @@ func (p *parser) Parse(ctx context.Context) (CpuStats, error) {
 	}
 }
 
-// parseForDarwin parses the CPU statistics of the system for Darwin
-func (p *parser) parseForDarwin(ctx context.Context) (CpuStats, error) {
-	// Using -l 1 for a single snapshot
-	cmd, args := cmdAndArgs(os.Darwin)
-	cmdRes, err := p.execer.Exec(cmd, args...)
-	if err != nil {
-		return CpuStats{}, err
-	}
-
-	res := CpuStats{}
-	for _, line := range cmdRes.Lines() {
-		if strings.Contains(line, "CPU usage:") {
-			parts := strings.Fields(line)
-			if len(parts) > 6 {
-				res.User, _ = strconv.ParseFloat(strings.TrimSuffix(parts[2], "%"), 64)
-				res.System, _ = strconv.ParseFloat(strings.TrimSuffix(parts[4], "%"), 64)
-				res.Idle, _ = strconv.ParseFloat(strings.TrimSuffix(parts[6], "%"), 64)
-			}
-			break
-		}
-	}
-
-	return res, nil
-}
-
 func cmdAndArgs(osystem string) (string, []string) {
 	switch osystem {
 	case os.Darwin:
@@ -104,42 +77,4 @@ func cmdAndArgs(osystem string) (string, []string) {
 	}
 
 	return "", nil
-}
-
-// parseForLinux parses the CPU statistics of the system for Linux
-func (p *parser) parseForLinux(ctx context.Context) (CpuStats, error) {
-	// Using -b -n 1 for batch mode and a single snapshot
-	cmdRes, err := p.execer.Exec("top", "-b", "-n", "1")
-	if err != nil {
-		return CpuStats{}, err
-	}
-
-	res := CpuStats{}
-	for _, line := range cmdRes.Lines() {
-		if strings.HasPrefix(line, "%Cpu(s):") {
-			parts := strings.Fields(line)
-			res.User, _ = strconv.ParseFloat(parts[1], 64)
-			res.System, _ = strconv.ParseFloat(parts[3], 64)
-			res.Idle, _ = strconv.ParseFloat(parts[7], 64)
-			break
-		}
-	}
-
-	return res, nil
-}
-
-// parseForWindows parses the CPU statistics of the system for Windows
-func (p *parser) parseForWindows(ctx context.Context) (CpuStats, error) {
-	// For Windows, wmic returns only CPU load
-	cmdRes, err := p.execer.Exec("wmic", "cpu", "get", "loadpercentage")
-	if err != nil {
-		return CpuStats{}, err
-	}
-
-	res := CpuStats{}
-	res.User, _ = strconv.ParseFloat(strings.TrimSpace(cmdRes.Lines()[0]), 64)
-	res.System = 0            // We don't get system load
-	res.Idle = 100 - res.User // idle is calculated as 100% - load
-
-	return res, nil
 }
